game object now has map calibration work in it
