+++ src/Maps/TestMap.java
@@ -15,7 +15,7 @@ import java.util.ArrayList;
 public class TestMap extends Map {
 
     public TestMap() {
-        super("test_map.txt", new CommonTileset(), new Point(20, 1));
+        super("test_map.txt", new CommonTileset(), new Point(1, 9));
     }
 
     @Override
